==> R/04_process_cps.R <==
# TODO: Implement CPS processing if needed.
# This script is optional and should process CPS data similarly to ACS processing.
# Key variables: EARNWEEK, UHRSWORKT, CPSOCCWT.
# Output should go to data/tidy/cps/
